2004-06-21  Laurent Sansonetti  <lrz@gnome.org>
        * lib/alexandria/book.rb:

        Splitted.

        * lib/alexandria/book_providers/amazon.rb,
          lib/alexandria/book_providers/proxis.rb,
          lib/alexandria/book_providers.rb,
          lib/alexandria/book_providers,
          lib/alexandria/library.rb:

        - Added ;
        - The Proxis provider has been contributed by Pascal Terjan ;
        - Do not save covers that are blank ;
        - The book providers factory should handle multiple providers now.

        * lib/alexandria/ui/icons.rb:

        Uses the default icon when the cover files don't exist.

        * lib/alexandria/ui/preferences_dialog.rb:

        Handles providers that don't provide any preference, and display
        an error message when the user tries to setup them.


git-svn-id: http://alexandria.rubyforge.org/svn/trunk/alexandria@145 9992d806-23ff-4cc5-b4c6-abbc17c6a14e

+++ lib/alexandria/ui/preferences_dialog.rb
@@ -71,6 +71,9 @@ module UI
     end
 
     class PreferencesDialog < GladeBase
+        include GetText
+        GetText.bindtextdomain(Alexandria::TEXTDOMAIN, nil, nil, "UTF-8")
+
         def initialize(parent, &changed_block)
             super('preferences_dialog.glade')
             @preferences_dialog.transient_for = parent
@@ -102,11 +105,14 @@ module UI
         end
 
         def on_setup
-            iter = @treeview_providers.selection.selected
-            provider = BookProviders.find { |x| x.name == iter[0] }
-            dialog = ProviderPreferencesDialog.new(@preferences_dialog, provider)
-            dialog.show_all.run
-            dialog.destroy
+            provider = selected_provider
+            if provider.prefs.empty?
+                ErrorDialog.new(@preferences_dialog, _("The '%s' provider doesn't have any preference to setup") % provider.name)
+            else
+                dialog = ProviderPreferencesDialog.new(@preferences_dialog, provider)
+                dialog.show_all.run
+                dialog.destroy
+            end
         end
 
         def on_column_toggled(checkbutton)
@@ -126,6 +132,15 @@ module UI
         
         def on_close
             @preferences_dialog.destroy
+        end
+
+        #######
+        private
+        #######
+
+        def selected_provider
+            iter = @treeview_providers.selection.selected
+            provider = BookProviders.find { |x| x.name == iter[0] }
         end
     end
 end
